chore: Fix linter findings for Windows (part4)
