+++ agent/agent_windows.go
@@ -4,10 +4,10 @@ package agent
 
 import "os"
 
-func watchForFlushSignal(flushRequested chan os.Signal) {
+func watchForFlushSignal(_ chan os.Signal) {
 	// not supported
 }
 
-func stopListeningForFlushSignal(flushRequested chan os.Signal) {
+func stopListeningForFlushSignal(_ chan os.Signal) {
 	// not supported
 }
