+++ plugins/inputs/diskio/sample.conf
@@ -1,5 +1,4 @@
 # Read metrics about disk IO by device
-# This plugin ONLY supports Linux
 [[inputs.diskio]]
   ## By default, telegraf will gather stats for all devices including
   ## disk partitions.
