+++ plugins/inputs/ethtool/ethtool_notlinux.go
@@ -3,9 +3,24 @@
 package ethtool
 
 import (
+	_ "embed"
+
 	"github.com/influxdata/telegraf"
 	"github.com/influxdata/telegraf/plugins/inputs"
 )
+
+//go:embed sample.conf
+var sampleConfig string
+
+const pluginName = "ethtool"
+
+type Ethtool struct {
+	Log telegraf.Logger `toml:"-"`
+}
+
+func (*Ethtool) SampleConfig() string {
+	return sampleConfig
+}
 
 func (e *Ethtool) Init() error {
 	e.Log.Warn("Current platform is not supported")
