+++ plugins/inputs/ping/ping_notwindows.go
@@ -15,6 +15,20 @@ import (
 	"github.com/influxdata/telegraf"
 )
 
+type roundTripTimeStats struct {
+	min    float64
+	avg    float64
+	max    float64
+	stddev float64
+}
+
+type statistics struct {
+	packetsTransmitted int
+	packetsReceived    int
+	ttl                int
+	roundTripTimeStats
+}
+
 func (p *Ping) pingToURL(u string, acc telegraf.Accumulator) {
 	tags := map[string]string{"url": u}
 	fields := map[string]interface{}{"result_code": 0}
@@ -67,11 +81,11 @@ func (p *Ping) pingToURL(u string, acc telegraf.Accumulator) {
 	}
 
 	// Calculate packet loss percentage
-	loss := float64(stats.trans-stats.recv) / float64(stats.trans) * 100.0
+	percentPacketLoss := float64(stats.packetsTransmitted-stats.packetsReceived) / float64(stats.packetsTransmitted) * 100.0
 
-	fields["packets_transmitted"] = stats.trans
-	fields["packets_received"] = stats.recv
-	fields["percent_packet_loss"] = loss
+	fields["packets_transmitted"] = stats.packetsTransmitted
+	fields["packets_received"] = stats.packetsReceived
+	fields["percent_packet_loss"] = percentPacketLoss
 	if stats.ttl >= 0 {
 		fields["ttl"] = stats.ttl
 	}
@@ -165,11 +179,11 @@ func (p *Ping) args(url string, system string) []string {
 //	round-trip min/avg/max/stddev = 34.843/43.508/52.172/8.664 ms
 //
 // It returns (<transmitted packets>, <received packets>, <average response>)
-func processPingOutput(out string) (stats, error) {
-	stats := stats{
-		trans: 0,
-		recv:  0,
-		ttl:   -1,
+func processPingOutput(out string) (statistics, error) {
+	stats := statistics{
+		packetsTransmitted: 0,
+		packetsReceived:    0,
+		ttl:                -1,
 		roundTripTimeStats: roundTripTimeStats{
 			min:    -1.0,
 			avg:    -1.0,
@@ -186,7 +200,7 @@ func processPingOutput(out string) (stats, error) {
 		if stats.ttl == -1 && (strings.Contains(line, "ttl=") || strings.Contains(line, "hlim=")) {
 			stats.ttl, err = getTTL(line)
 		} else if strings.Contains(line, "transmitted") && strings.Contains(line, "received") {
-			stats.trans, stats.recv, err = getPacketStats(line)
+			stats.packetsTransmitted, stats.packetsReceived, err = getPacketStats(line)
 			if err != nil {
 				return stats, err
 			}
