+++ plugins/inputs/ping/ping_test.go
@@ -82,8 +82,8 @@ func TestProcessPingOutput(t *testing.T) {
 	stats, err := processPingOutput(bsdPingOutput)
 	require.NoError(t, err)
 	require.Equal(t, 55, stats.ttl, "ttl value is 55")
-	require.Equal(t, 5, stats.trans, "5 packets were transmitted")
-	require.Equal(t, 5, stats.recv, "5 packets were received")
+	require.Equal(t, 5, stats.packetsTransmitted, "5 packets were transmitted")
+	require.Equal(t, 5, stats.packetsReceived, "5 packets were received")
 	require.InDelta(t, 15.087, stats.min, 0.001)
 	require.InDelta(t, 20.224, stats.avg, 0.001)
 	require.InDelta(t, 27.263, stats.max, 0.001)
@@ -92,8 +92,8 @@ func TestProcessPingOutput(t *testing.T) {
 	stats, err = processPingOutput(freebsdPing6Output)
 	require.NoError(t, err)
 	require.Equal(t, 117, stats.ttl, "ttl value is 117")
-	require.Equal(t, 5, stats.trans, "5 packets were transmitted")
-	require.Equal(t, 5, stats.recv, "5 packets were received")
+	require.Equal(t, 5, stats.packetsTransmitted, "5 packets were transmitted")
+	require.Equal(t, 5, stats.packetsReceived, "5 packets were received")
 	require.InDelta(t, 35.727, stats.min, 0.001)
 	require.InDelta(t, 53.211, stats.avg, 0.001)
 	require.InDelta(t, 93.870, stats.max, 0.001)
@@ -102,8 +102,8 @@ func TestProcessPingOutput(t *testing.T) {
 	stats, err = processPingOutput(linuxPingOutput)
 	require.NoError(t, err)
 	require.Equal(t, 63, stats.ttl, "ttl value is 63")
-	require.Equal(t, 5, stats.trans, "5 packets were transmitted")
-	require.Equal(t, 5, stats.recv, "5 packets were received")
+	require.Equal(t, 5, stats.packetsTransmitted, "5 packets were transmitted")
+	require.Equal(t, 5, stats.packetsReceived, "5 packets were received")
 	require.InDelta(t, 35.225, stats.min, 0.001)
 	require.InDelta(t, 43.628, stats.avg, 0.001)
 	require.InDelta(t, 51.806, stats.max, 0.001)
@@ -112,8 +112,8 @@ func TestProcessPingOutput(t *testing.T) {
 	stats, err = processPingOutput(busyBoxPingOutput)
 	require.NoError(t, err)
 	require.Equal(t, 56, stats.ttl, "ttl value is 56")
-	require.Equal(t, 4, stats.trans, "4 packets were transmitted")
-	require.Equal(t, 4, stats.recv, "4 packets were received")
+	require.Equal(t, 4, stats.packetsTransmitted, "4 packets were transmitted")
+	require.Equal(t, 4, stats.packetsReceived, "4 packets were received")
 	require.InDelta(t, 15.810, stats.min, 0.001)
 	require.InDelta(t, 17.611, stats.avg, 0.001)
 	require.InDelta(t, 22.559, stats.max, 0.001)
@@ -139,8 +139,8 @@ func TestProcessPingOutputWithVaryingTTL(t *testing.T) {
 	stats, err := processPingOutput(linuxPingOutputWithVaryingTTL)
 	require.NoError(t, err)
 	require.Equal(t, 63, stats.ttl, "ttl value is 63")
-	require.Equal(t, 5, stats.trans, "5 packets were transmitted")
-	require.Equal(t, 5, stats.recv, "5 packets were transmitted")
+	require.Equal(t, 5, stats.packetsTransmitted, "5 packets were transmitted")
+	require.Equal(t, 5, stats.packetsReceived, "5 packets were transmitted")
 	require.InDelta(t, 35.225, stats.min, 0.001)
 	require.InDelta(t, 43.628, stats.avg, 0.001)
 	require.InDelta(t, 51.806, stats.max, 0.001)
