--- plugins/inputs/postfix/stat_none.go
@@ -1,11 +1,0 @@
-//go:build !dragonfly && !linux && !netbsd && !openbsd && !solaris && !darwin && !freebsd
-
-package postfix
-
-import (
-	"time"
-)
-
-func statCTime(_ interface{}) time.Time {
-	return time.Time{}
-}
