+++ plugins/inputs/processes/processes_windows.go
@@ -16,7 +16,7 @@ func (e *Processes) Init() error {
 	return nil
 }
 
-func (e *Processes) Gather(acc telegraf.Accumulator) error {
+func (e *Processes) Gather(_ telegraf.Accumulator) error {
 	return nil
 }
